renaming in package com.playonlinux.core.injection in com.playonlinux.injection
