+++ phoenicis-app/src/main/java/com/playonlinux/app/PlayOnLinuxApp.java
@@ -18,10 +18,10 @@
 
 package com.playonlinux.app;
 
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.InjectionException;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.core.services.manager.ServiceManager;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.InjectionException;
+import com.playonlinux.injection.Scan;
 import com.playonlinux.ui.api.Controller;
 
 /**
