+++ phoenicis-app/src/main/java/com/playonlinux/app/PlayOnLinuxConfig.java
@@ -27,8 +27,6 @@ import java.util.concurrent.Executors;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.playonlinux.containers.AnyContainerFactory;
 import com.playonlinux.containers.WinePrefixContainerFactory;
-import com.playonlinux.core.injection.AbstractConfiguration;
-import com.playonlinux.core.injection.Bean;
 import com.playonlinux.core.lang.LanguageBundle;
 import com.playonlinux.core.lang.LanguageBundleSelector;
 import com.playonlinux.core.log.LoggerFactory;
@@ -40,6 +38,8 @@ import com.playonlinux.core.services.manager.PlayOnLinuxServicesManager;
 import com.playonlinux.core.services.manager.ServiceManager;
 import com.playonlinux.engines.wine.WineVersionSource;
 import com.playonlinux.engines.wine.WineversionsSourceWebserviceDefaultImplementation;
+import com.playonlinux.injection.AbstractConfiguration;
+import com.playonlinux.injection.Bean;
 import com.playonlinux.ui.api.Controller;
 import com.playonlinux.ui.events.EventHandler;
 import com.playonlinux.ui.events.EventHandlerPlayOnLinuxImplementation;
