+++ phoenicis-core/src/main/java/com/playonlinux/core/observer/ObservableDirectory.java
@@ -22,10 +22,10 @@ import java.io.File;
 
 import org.apache.commons.lang.builder.ToStringBuilder;
 
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.core.services.manager.Service;
 import com.playonlinux.core.services.manager.ServiceManager;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 
 @Scan
 public abstract class ObservableDirectory<T> extends ObservableDefaultImplementation<T> implements Service, AutoCloseable{
