+++ phoenicis-core/src/main/java/com/playonlinux/core/python/JythonCommandInterpreter.java
@@ -28,10 +28,10 @@ import org.python.core.PyException;
 import org.python.util.InteractiveInterpreter;
 
 import com.playonlinux.app.PlayOnLinuxException;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.core.services.manager.Service;
 import com.playonlinux.core.services.manager.ServiceManager;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 
 @Scan
 public class JythonCommandInterpreter implements CommandInterpreter, Service {
