+++ phoenicis-core/src/main/java/com/playonlinux/framework/Environment.java
@@ -23,12 +23,12 @@ import static com.playonlinux.core.EnvironmentUtils.mergeEnvironmentVariables;
 import java.util.Map;
 
 import com.playonlinux.app.PlayOnLinuxContext;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.core.scripts.ScriptClass;
 import com.playonlinux.core.scripts.ScriptFailureException;
 import com.playonlinux.core.utils.Architecture;
 import com.playonlinux.core.utils.OperatingSystem;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 
 @ScriptClass
 @Scan
