+++ phoenicis-core/src/main/java/com/playonlinux/library/LibraryEventHandler.java
@@ -24,9 +24,9 @@ import java.io.IOException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.playonlinux.app.PlayOnLinuxContext;
 import com.playonlinux.app.PlayOnLinuxException;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.framework.wizard.WineWizard;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 import com.playonlinux.library.shortcuts.Shortcut;
 
 @Scan
