+++ phoenicis-core/src/main/java/com/playonlinux/library/ShortcutCreator.java
@@ -22,8 +22,8 @@ import java.io.File;
 import java.io.IOException;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 import com.playonlinux.library.shortcuts.Shortcut;
 
 @Scan
