+++ phoenicis-core/src/main/java/com/playonlinux/ui/events/EventHandlerPlayOnLinuxImplementation.java
@@ -30,14 +30,14 @@ import com.playonlinux.apps.InstallerDownloaderEntityProvider;
 import com.playonlinux.containers.ContainersEntitiesProvider;
 import com.playonlinux.containers.entities.ContainerEntity;
 import com.playonlinux.containers.entities.ContainersWindowEntity;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.core.scripts.AnyScriptFactory;
 import com.playonlinux.core.scripts.Script;
 import com.playonlinux.core.services.manager.ServiceManager;
 import com.playonlinux.engines.wine.WineVersionEntitiesProvider;
 import com.playonlinux.engines.wine.entities.WineVersionDistributionItemEntity;
 import com.playonlinux.engines.wine.entities.WineVersionsWindowEntity;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 import com.playonlinux.library.LibraryEntitiesProvider;
 import com.playonlinux.library.LibraryEventHandler;
 import com.playonlinux.library.entities.InstalledApplicationEntity;
