+++ phoenicis-core/src/test/java/com/playonlinux/framework/SetupWizardTest.java
@@ -31,11 +31,11 @@ import org.junit.Test;
 
 import com.playonlinux.app.PlayOnLinuxContext;
 import com.playonlinux.app.PlayOnLinuxException;
-import com.playonlinux.core.injection.AbstractConfiguration;
-import com.playonlinux.core.injection.Bean;
-import com.playonlinux.core.injection.InjectionException;
 import com.playonlinux.core.lang.FallbackLanguageBundle;
 import com.playonlinux.core.lang.LanguageBundle;
+import com.playonlinux.injection.AbstractConfiguration;
+import com.playonlinux.injection.Bean;
+import com.playonlinux.injection.InjectionException;
 import com.playonlinux.ui.api.Controller;
 import com.playonlinux.ui.impl.mockui.MockUIMessageSenderImplementation;
 import com.playonlinux.ui.impl.mockui.setupwindow.MockUISetupWindowImplementation;
