+++ phoenicis-injection/src/main/java/com/playonlinux/injection/AbstractConfiguration.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package com.playonlinux.core.injection;
+package com.playonlinux.injection;
 
 import java.util.Map;
 import java.util.concurrent.Semaphore;
