+++ phoenicis-injection/src/main/java/com/playonlinux/injection/InjectionException.java
@@ -16,14 +16,18 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package com.playonlinux.core.injection;
+package com.playonlinux.injection;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+public class InjectionException extends Exception {
+    private final String error;
 
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Scan {
+    public InjectionException(String error, Throwable e) {
+        this.error = error;
+        this.initCause(e);
+    }
+
+    @Override
+    public String toString() {
+        return this.error;
+    }
 }
