+++ phoenicis-javafx/src/main/java/com/playonlinux/ui/impl/javafx/mainwindow/containers/EventHandlerContainers.java
@@ -20,8 +20,8 @@ package com.playonlinux.ui.impl.javafx.mainwindow.containers;
 
 import com.playonlinux.containers.entities.ContainerEntity;
 import com.playonlinux.containers.entities.ContainersWindowEntity;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 import com.playonlinux.ui.api.EntitiesProvider;
 import com.playonlinux.ui.events.EventHandler;
