+++ phoenicis-wine/src/main/java/com/playonlinux/framework/WineShortcut.java
@@ -22,12 +22,12 @@ import java.io.File;
 import java.util.List;
 
 import com.playonlinux.app.PlayOnLinuxContext;
-import com.playonlinux.core.injection.Inject;
-import com.playonlinux.core.injection.Scan;
 import com.playonlinux.core.scripts.CancelException;
 import com.playonlinux.core.scripts.ScriptClass;
 import com.playonlinux.core.scripts.ScriptFailureException;
 import com.playonlinux.framework.wizard.SetupWizardComponent;
+import com.playonlinux.injection.Inject;
+import com.playonlinux.injection.Scan;
 import com.playonlinux.library.LibraryException;
 import com.playonlinux.library.ShortcutCreator;
